Issue #462 Split up the conversion content of the metadataVariantHelper and its unit tests.
