+++ bouncer/src/repo/manipulator/modelconvertor/import/odaHelper/data_processor_rvt.cpp
@@ -42,6 +42,120 @@ const std::set<std::string> IGNORE_PARAMS = {
 	"RENDER APPEARANCE PROPERTIES"
 };
 
+bool repo::manipulator::modelconvertor::odaHelper::TryConvertMetadataEntry(OdTfVariant& metaEntry, OdBmLabelUtilsPEPtr labelUtils, OdBmParamDefPtr paramDef, OdBmDatabase* database, OdBm::BuiltInParameter::Enum param, repo::lib::MetadataVariant& v)
+{
+	switch (metaEntry.type()) {
+	case OdVariant::kVoid: {
+		return false;
+	}
+	case OdVariant::kString: {
+		v = repo::manipulator::modelconvertor::odaHelper::convertToStdString(metaEntry.getString());
+		break;
+	}
+	case OdVariant::kBool:
+	{
+		v = metaEntry.getBool();
+		break;
+	}
+	case OdVariant::kInt8: {
+		OdInt8 value = metaEntry.getInt8();
+		v = static_cast<int>(value);
+		break;
+	}
+	case OdVariant::kInt16: {
+		OdInt16 value = metaEntry.getInt16();
+		v = static_cast<int>(value);
+		break;
+	}
+	case OdVariant::kInt32: {
+		if (paramDef->getParameterTypeId() == OdBmSpecTypeId::Boolean::kYesNo)
+		{
+			(metaEntry.getInt32()) ? v = true : v = false;
+		}
+		else
+		{
+			OdInt32 value = metaEntry.getInt32();
+			v = static_cast<long long>(value);
+		}
+		break;
+	}
+	case OdVariant::kInt64: {
+		OdInt64 value = metaEntry.getInt64();
+		v = static_cast<long long>(value);
+		break;
+	}
+	case OdVariant::kDouble: {
+		OdDouble value = metaEntry.getDouble();
+		v = static_cast<double>(value);
+		break;
+	}
+						   break;
+	case OdVariant::kAnsiString: {
+		v = std::string(metaEntry.getAnsiString().c_str());
+		break;
+	}
+	case OdTfVariant::kDbStubPtr: {
+		// A stub is effectively a pointer to another database, or built-in, object. We don't recurse these objects, but will try to extract
+		// their names or identities where possible (e.g. if a key pointed to a wall object, we would get the name of the wall object).
+
+		OdDbStub* stub = metaEntry.getDbStubPtr();
+		OdBmObjectId bmId = OdBmObjectId(stub);
+		if (stub)
+		{
+			OdDbHandle hdl = bmId.getHandle();
+			if (param == OdBm::BuiltInParameter::ELEM_CATEGORY_PARAM || param == OdBm::BuiltInParameter::ELEM_CATEGORY_PARAM_MT)
+			{
+				if (OdBmObjectId::isRegularHandle(hdl)) // A regular handle points to a database entry; if it is not regular, it is built-in.
+				{
+					v = std::to_string((OdUInt64)hdl);
+				}
+				else
+				{
+					OdBm::BuiltInCategory::Enum builtInValue = static_cast<OdBm::BuiltInCategory::Enum>((OdUInt64)hdl);
+					auto category = labelUtils->getLabelFor(OdBm::BuiltInCategory::Enum(builtInValue));
+					if (!category.isEmpty()) {
+						v = repo::manipulator::modelconvertor::odaHelper::convertToStdString(category);
+					}
+					else {
+						v = repo::manipulator::modelconvertor::odaHelper::convertToStdString(OdBm::BuiltInCategory(builtInValue).toString());
+					}
+				}
+			}
+			else
+			{
+				OdBmObjectPtr bmPtr = bmId.openObject();
+				if (!bmPtr.isNull())
+				{
+					// The object class is unknown - some superclasses we can handle explicitly.
+
+					auto bmPtrClass = bmPtr->isA();
+					if (!bmPtrClass->isKindOf(OdBmElement::desc()))
+					{
+						OdBmElementPtr elem = bmPtr;
+						if (elem->getElementName() == OdString::kEmpty) {
+							v = std::to_string((OdUInt64)bmId.getHandle());
+						}
+						else {
+							v = repo::manipulator::modelconvertor::odaHelper::convertToStdString(elem->getElementName());
+						}
+					}
+					else
+					{
+						repoError << "Unsupported metadata value type (class) " << repo::manipulator::modelconvertor::odaHelper::convertToStdString(bmPtrClass->name()) << " currently only OdBmElement's are supported.";
+					}
+				}
+			}
+		}
+		break;
+	}
+	default:
+		return false;
+	}
+
+	return true;
+}
+
+
 bool DataProcessorRvt::ignoreParam(const std::string& param)
 {
 	std::string paramUpper = param; // Copy the string to transform it in place
@@ -335,7 +449,7 @@ void DataProcessorRvt::processParameter(
 			repo::lib::MetadataVariant v;
 
 
-			if (repo::lib::MetadataVariantHelper::TryConvert(value, labelUtils, pDescParam, element->getDatabase(), buildInEnum, v))
+			if (TryConvertMetadataEntry(value, labelUtils, pDescParam, element->getDatabase(), buildInEnum, v))
 			{
 				if (metadata.find(metaKey) != metadata.end() && !boost::apply_visitor(repo::lib::DuplicationVisitor(), metadata[metaKey], v)) {
