+++ test/src/unit/repo/lib/CMakeLists.txt
@@ -6,8 +6,12 @@
 
 set(TEST_SOURCES
 	${TEST_SOURCES}
+	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_MetadataVariant.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_config.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_matrix.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_metadata_variant_assimp.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_metadata_variant_nwd.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_metadata_variant_rvt.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_uuid.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_vector2d.cpp
 	CACHE STRING "TEST_SOURCES" FORCE)
