+++ test/src/unit/repo/lib/ut_repo_MetadataVariant.cpp
@@ -15,6 +15,7 @@
 *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+#include "repo/manipulator/modelconvertor/import/odaHelper/helper_functions.h"
 #include <cstdlib>
 #include <repo/lib/datastructure/repo_metadataVariant.h>
 #include <repo/lib/datastructure/repo_metadataVariantHelper.h>
@@ -22,27 +23,6 @@
 
 #include "../../repo_test_utils.h"
 
-#include <Attribute/NwPropertyAttribute.h>
-#include <NwVariant.h>
-#include <NwDatabase.h>
-#include <StaticRxObject.h>
-#include <ExSystemServices.h>
-#include <NwHostAppServices.h>
-#include "DynamicLinker.h"
-#include "Gs/GsBaseModule.h"
-
-#include "StaticRxObject.h"
-#include "ExBimHostAppServices.h"
-#include <Database/BmParamDefHelper.h>
-#include "HostObj/Entities/BmSWall.h"
-#include "Database/BmDatabase.h"
-//#include <Database/BmTransaction.h>
-#include "TB_Commands/TBCommandsStdAfx.h"
-
-#include "Essential/Entities/BmExtrusionElem.h"
-#include "Database/Entities/BmDirectShape.h"
-
-#include "Essential/Entities/BmBlendElem.h"
 
 using namespace repo::lib;
 
@@ -197,944 +177,3 @@ TEST(RepoMetaVariantTest, CompareVisitor) {
 
 	EXPECT_FALSE(boost::apply_visitor(DuplicationVisitor(), v6a, v6b));
 }
-
-TEST(RepoMetaVariantConverterAssimpTest, Boolean)
-{
-	bool data = true;
-	aiMetadataEntry meta;
-	meta.mType = aiMetadataType::AI_BOOL;
-	meta.mData = &data;
-
-	MetadataVariant v;
-	EXPECT_TRUE(MetadataVariantHelper::TryConvert(meta, v));
-	EXPECT_EQ(boost::get<bool>(v), true);
-}
-
-TEST(RepoMetaVariantConverterAssimpTest, Int32)
-{
-	// Int32 (positive)
-	int dataPos = 2147483647;
-	aiMetadataEntry metaPos;
-	metaPos.mType = aiMetadataType::AI_INT32;
-	metaPos.mData = &dataPos;
-
-	MetadataVariant vPos;
-	EXPECT_TRUE(MetadataVariantHelper::TryConvert(metaPos, vPos));
-	EXPECT_EQ(boost::get<int>(vPos), 2147483647);
-
-	// Int32 (negative)
-	int dataNeg = -2147483647;
-	aiMetadataEntry metaNeg;
-	metaNeg.mType = aiMetadataType::AI_INT32;
-	metaNeg.mData = &dataNeg;
-
-	MetadataVariant vNeg;
-	EXPECT_TRUE(MetadataVariantHelper::TryConvert(metaNeg, vNeg));
-	EXPECT_EQ(boost::get<int>(vNeg), -2147483647);
-}
-
-
-TEST(RepoMetaVariantConverterAssimpTest, UInt64)
-{
-	// UInt64 (min)
-	uint64_t dataMin = 0;
-	aiMetadataEntry metaMin;
-	metaMin.mType = aiMetadataType::AI_UINT64;
-	metaMin.mData = &dataMin;
-
-	MetadataVariant vMin;
-	EXPECT_TRUE(MetadataVariantHelper::TryConvert(metaMin, vMin));
-	EXPECT_EQ(boost::get<long long>(vMin), 0);
-
-	// Uint64 (max)
-	uint64_t dataMax = 18446744073709551615;
-	aiMetadataEntry metaMax;
-	metaMax.mType = aiMetadataType::AI_UINT64;
-	metaMax.mData = &dataMax;
-
-	MetadataVariant vMax;
-	EXPECT_TRUE(MetadataVariantHelper::TryConvert(metaMax, vMax));
-	EXPECT_EQ(boost::get<long long>(vMax), 18446744073709551615);
-}
-
-TEST(RepoMetaVariantConverterAssimpTest, Float)
-{
-	float data = 0.5;
-	aiMetadataEntry meta;
-	meta.mType = aiMetadataType::AI_FLOAT;
-	meta.mData = &data;
-
-	MetadataVariant v;
-	EXPECT_TRUE(MetadataVariantHelper::TryConvert(meta, v));
-	EXPECT_EQ(boost::get<double>(v), 0.5);
-}
-
-TEST(RepoMetaVariantConverterAssimpTest, Double)
-{
-	double data = 0.6;
-	aiMetadataEntry meta;
-	meta.mType = aiMetadataType::AI_DOUBLE;
-	meta.mData = &data;
-
-	MetadataVariant v;
-	EXPECT_TRUE(MetadataVariantHelper::TryConvert(meta, v));
-	EXPECT_EQ(boost::get<double>(v), 0.6);
-}
-
-TEST(RepoMetaVariantConverterAssimpTest, Vector)
-{
-	aiVector3D data;
-	data.x = 1.00;
-	data.y = 2.00;
-	data.z = 3.00;
-	aiMetadataEntry meta;
-	meta.mType = aiMetadataType::AI_AIVECTOR3D;
-	meta.mData = &data;
-
-	MetadataVariant v;
-	EXPECT_TRUE(MetadataVariantHelper::TryConvert(meta, v));
-	EXPECT_EQ(boost::get<std::string>(v), "[1.00000000000000000, 2.00000000000000000, 3.00000000000000000]");
-}
-
-TEST(RepoMetaVariantConverterAssimpTest, AISTRING)
-{
-	aiString data = aiString("test");
-	aiMetadataEntry meta;
-	meta.mType = aiMetadataType::AI_AISTRING;
-	meta.mData = &data;
-
-	MetadataVariant v;
-	EXPECT_FALSE(MetadataVariantHelper::TryConvert(meta, v));
-}
-
-
-// First helper class for the test of the NWD Converter
-class OdExNwSystemServices : public ExSystemServices
-{
-public:
-	OdExNwSystemServices() {}
-};
-
-
-// Second helper class for the test of the NWD Converter
-class MyNwServices : public OdExNwSystemServices, public OdNwHostAppServices {
-protected:
-	ODRX_USING_HEAP_OPERATORS(OdExNwSystemServices);
-};
-
-void SetupOdEnvForNWD(OdStaticRxObject<MyNwServices> &svcs, OdNwDatabasePtr &pDB) {
-	odrxInitialize(&svcs);
-	odgsInitialize();
-	::odrxDynamicLinker()->loadModule(OdNwDbModuleName, false);
-
-	pDB = svcs.createDatabase(true);
-}
-
-void TeardownOdEnvForNWD() {
-	::odrxDynamicLinker()->unloadUnreferenced();
-	odgsUninitialize();
-	::odrxUninitialize();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, DoubleTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	double value = 1.0;
-	OdNwVariant var = OdNwVariant(value);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("double"), OdString("double"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<double>(v), 1.0);
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, Int32Test)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	OdInt32 value = -2147483648ll;
-	OdNwVariant var = OdNwVariant(value);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("int32"), OdString("int32"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<long long>(v), -2147483648ll);
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, BoolTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	bool value = true;
-	OdNwVariant var = OdNwVariant(value);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("bool"), OdString("bool"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<bool>(v), true);
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, DisplayStringTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	OdString value = OdString("Test");
-	OdNwVariant var = OdNwVariant(value);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("dispString"), OdString("dispString"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<std::string>(v), std::string("Test"));
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, TimeTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	OdUInt64 value = 781747200ll;
-	OdNwVariant var = OdNwVariant(value);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("time"), OdString("time"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(mktime(&boost::get<tm>(v)), value);
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, LengthTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	double value = 2.0;
-	OdNwVariant var = OdNwVariant(value, OdNwVariant::Type::kLength);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("length"), OdString("length"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<double>(v), 2.0);
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, AngleTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	double value = 3.0;
-	OdNwVariant var = OdNwVariant(value, OdNwVariant::Type::kAngle);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("angle"), OdString("angle"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<double>(v), 3.0);
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, NameTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	OdNwNamePtr value = OdNwName::createObject();
-	OdNwVariant var = OdNwVariant(OdRxObject::cast(value));
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("name"), OdString("name"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<std::string>(v), std::string("")); // Not ideal, but I have not found a way to set the OdNwName object yet (FT).
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, IDStringTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	OdString value = OdString("ID");
-	OdNwVariant var = OdNwVariant(value);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("idStr"), OdString("idStr"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<std::string>(v), std::string("ID"));
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, AreaTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	double value = 4.0;
-	OdNwVariant var = OdNwVariant(value, OdNwVariant::Type::kArea);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("area"), OdString("area"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<double>(v), 4.0);
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, VolumeTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	double value = 5.0;
-	OdNwVariant var = OdNwVariant(value, OdNwVariant::Type::kVolume);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("volume"), OdString("volume"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<double>(v), 5.0);
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, Point2DTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	OdGePoint2d value = OdGePoint2d(1.0, 2.0);
-	OdNwVariant var = OdNwVariant(value);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("point2D"), OdString("point2D"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<std::string>(v), std::string("1.000000, 2.000000"));
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-TEST(RepoMetaVariantConverterNWDTest, Point3DTest)
-{
-	// Setup of the OD environment
-	OdStaticRxObject<MyNwServices> svcs;
-	OdNwDatabasePtr pDB;
-	SetupOdEnvForNWD(svcs, pDB);
-
-	// Create data
-	OdGePoint3d value = OdGePoint3d(1.0, 2.0, 3.0);
-	OdNwVariant var = OdNwVariant(value);
-
-	// Convert to property
-	OdNwPropertyAttributePtr attribute = OdNwPropertyAttribute::createObject();
-	attribute->addProperty(OdString("point3D"), OdString("point3D"), var);
-
-	// Retrieve property
-	OdArray<OdNwDataPropertyPtr> props;
-	attribute->getProperties(props);
-
-	// Convert to MetadataVariant
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(props[0], v);
-
-	// Check results
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<std::string>(v), std::string("1.000000, 2.000000, 3.000000"));
-
-	// Teardown
-	TeardownOdEnvForNWD();
-}
-
-// First helper class for the test of the RVT Converter
-class OdExBimSystemServices : public ExSystemServices
-{
-public:
-	OdExBimSystemServices() {}
-};
-
-
-// Second helper class for the test of the RVT Converter
-class MyRvServices : public OdExBimSystemServices, public OdExBimHostAppServices {
-protected:
-	ODRX_USING_HEAP_OPERATORS(OdExBimSystemServices);
-};
-
-void SetupOdEnvForRvt(OdStaticRxObject<MyRvServices>& svcs, OdBmDatabasePtr& pDB) {
-	odrxInitialize(&svcs);
-	odgsInitialize();
-	::odrxDynamicLinker()->loadModule(OdBmLoaderModuleName, false);
-
-	pDB = svcs.createDatabase(true);
-}
-
-void TeardownOdEnvForRvt() {
-	::odrxDynamicLinker()->unloadUnreferenced();
-	odgsUninitialize();
-	::odrxUninitialize();
-}
-
-TEST(RepoMetaVariantConverterRevitTest, EmptyTest) {
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBmDatabase* database;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdTfVariant variant;	
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, database, param, v);
-
-	// Check result
-	EXPECT_FALSE(success);
-}
-
-TEST(RepoMetaVariantConverterRevitTest, StringTest) {
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBmDatabase* database;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdTfVariant variant = OdString("Test");
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, database, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<std::string>(v), std::string("Test"));
-}
-
-TEST(RepoMetaVariantConverterRevitTest, BoolTest) {
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBmDatabase* database;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdTfVariant variant = true;
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, database, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<bool>(v), true);
-}
-
-TEST(RepoMetaVariantConverterRevitTest, Int8Test) {
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBmDatabase* database;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdInt8 data = -128;
-	OdTfVariant variant = data;
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, database, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<int>(v), -128);
-}
-
-TEST(RepoMetaVariantConverterRevitTest, Int16Test) {
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBmDatabase* database;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdInt16 data = -32768;
-	OdTfVariant variant = data;
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, database, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<int>(v), -32768);
-}
-
-
-TEST(RepoMetaVariantConverterRevitTest, Int32AsBoolTest) {
-
-	// Setup of the OD environment
-	OdStaticRxObject<MyRvServices> svcs;
-	OdBmDatabasePtr pDB;
-	SetupOdEnvForRvt(svcs, pDB);
-
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Configure paramDef
-	paramDef = OdBmParamDefHelper::createParamDef(OdBmSpecTypeId::Boolean::kYesNo, OdBm::StorageType::Integer);
-
-	// Create data
-	OdInt32 data = 0;
-	OdTfVariant variant = data;
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, pDB, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<bool>(v), false);
-
-	// Teardown
-	TeardownOdEnvForRvt();
-}
-
-TEST(RepoMetaVariantConverterRevitTest, Int32asIntTest) {
-
-	// Setup of the OD environment
-	OdStaticRxObject<MyRvServices> svcs;
-	OdBmDatabasePtr pDB;
-	SetupOdEnvForRvt(svcs, pDB);
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdInt32 data = -2147483648ll;
-	OdTfVariant variant = data;
-
-	// Configure paramDef (to anything but kYesNo)
-	paramDef = OdBmParamDefHelper::createParamDef(OdBmSpecTypeId::Int::kInteger, OdBm::StorageType::Integer);
-
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, pDB, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<long long>(v), -2147483648ll);
-
-	// Teardown
-	TeardownOdEnvForRvt();
-}
-
-TEST(RepoMetaVariantConverterRevitTest, Int64Test) {
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBmDatabase* database;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdInt64 data = -9223372036854775808ll;
-	OdTfVariant variant = data;
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, database, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<long long>(v), -9223372036854775808ll);
-}
-
-TEST(RepoMetaVariantConverterRevitTest, DoubleTest) {
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBmDatabase* database;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdDouble data = 1.0;
-	OdTfVariant variant = data;
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, database, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<double>(v), 1.0);
-}
-
-TEST(RepoMetaVariantConverterRevitTest, AnsiStringTest) {
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBmDatabase* database;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdAnsiString data = "Test";
-	OdTfVariant variant = data;
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, database, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	EXPECT_EQ(boost::get<std::string>(v), std::string("Test"));
-}
-
-//TEST(RepoMetaVariantRevitConverterTest, StubPtrDataTest) {
-//
-//	// Setup of the OD environment
-//	OdStaticRxObject<MyRvServices> svcs;
-//
-//	odrxInitialize(&svcs);
-//	odgsInitialize();
-//	::odrxDynamicLinker()->loadModule(OdBmLoaderModuleName, false);
-//
-//
-//
-//	OdBmDatabasePtr pDB;
-//	pDB = svcs.createDatabase(true);
-//
-//
-//
-//	// Set up
-//	OdBmLabelUtilsPEPtr labelUtils;
-//	OdBmParamDefPtr paramDef;
-//	OdBm::BuiltInParameter::Enum param;
-//
-//	// Create data
-//
-//	OdBmElementPtr elementPtr = OdBmElement::createObject();
-//	elementPtr->setElementName("Test");
-//	
-//	OdBmObjectPtr objectPtr = elementPtr;
-//
-//	auto objectId = elementPtr->objectId();
-//
-//	OdDbStub* stub(objectId);
-//	
-//
-//	OdTfVariant variant = stub;
-//
-//	// Configure paramDef (to anything but kYesNo)
-//	//paramDef = OdBmParamDefHelper::createParamDef(OdBmSpecTypeId::Int::kInteger, OdBm::StorageType::Integer);
-//
-//
-//	// Convert
-//	MetadataVariant v;
-//	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, pDB, param, v);
-//
-//	// Check result
-//	EXPECT_TRUE(success);
-//	EXPECT_EQ(boost::get<std::string>(v), std::string("test"));
-//
-//	// Teardown
-//	::odrxDynamicLinker()->unloadUnreferenced();
-//	odgsUninitialize();
-//	::odrxUninitialize();
-//}
-
-TEST(RepoMetaVariantConverterRevitTest, StubPtrDataTestNoParamNoName) {
-
-	// Setup of the OD environment
-	OdStaticRxObject<MyRvServices> svcs;
-
-	odrxInitialize(&svcs);
-	odgsInitialize();
-	::odrxDynamicLinker()->loadModule(OdBmLoaderModuleName, false);
-
-	OdBmDatabasePtr pDB;
-	pDB = svcs.createTemplate(OdBm::ProjectTemplate::Structural, OdBm::UnitSystem::Metric);
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBm::BuiltInParameter::Enum param;
-
-	// Create data
-	OdBmTransaction tr(pDB, OdBm::TransactionType::Regular);
-	tr.start();
-
-	OdBmObjectId id;	
-	OdResult result;
-	OdBmDirectShapePtr pDirectShape = OdBmDirectShape::createObject();
-
-	result = pDB->addElement(pDirectShape, id);
-
-	tr.commit();
-
-	OdDbStub* stub(id);
-	OdTfVariant variant = stub;
-
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, pDB, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	std::string variantResult = boost::get<std::string>(v);
-	EXPECT_EQ(variantResult, std::to_string((OdUInt64)id.getHandle()));
-
-}
-
-TEST(RepoMetaVariantConverterRevitTest, StubPtrDataTestParamSetRegularHdl) {
-
-	// Setup of the OD environment
-	OdStaticRxObject<MyRvServices> svcs;
-
-	odrxInitialize(&svcs);
-	odgsInitialize();
-	::odrxDynamicLinker()->loadModule(OdBmLoaderModuleName, false);
-
-	OdBmDatabasePtr pDB;
-	pDB = svcs.createTemplate(OdBm::ProjectTemplate::Structural, OdBm::UnitSystem::Metric);
-
-	// Set up
-	OdBmLabelUtilsPEPtr labelUtils;
-	OdBmParamDefPtr paramDef;
-	OdBm::BuiltInParameter::Enum param = OdBm::BuiltInParameter::ELEM_CATEGORY_PARAM;
-
-	// Create data
-	OdBmTransaction tr(pDB, OdBm::TransactionType::Regular);
-	tr.start();
-
-	OdBmObjectId id;
-	OdResult result;
-	OdBmDirectShapePtr pDirectShape = OdBmDirectShape::createObject();
-
-	result = pDB->addElement(pDirectShape, id);
-
-	tr.commit();
-
-	OdDbStub* stub(id);
-	OdTfVariant variant = stub;
-
-
-	// Convert
-	MetadataVariant v;
-	bool success = MetadataVariantHelper::TryConvert(variant, labelUtils, paramDef, pDB, param, v);
-
-	// Check result
-	EXPECT_TRUE(success);
-	std::string variantResult = boost::get<std::string>(v);
-	EXPECT_EQ(variantResult, std::to_string((OdUInt64)id.getHandle()));
-
-}
-
-
